+++ util/src/main/scala/org/broadinstitute/dsde/workbench/util/syntax.scala
@@ -1,0 +1,6 @@
+package org.broadinstitute.dsde.workbench.util
+import io.chrisdavenport.log4cats.Logger
+
+object syntax {
+  implicit def logSyntax[F[_]](log: Logger[F]): ContextLogger[F] = ContextLogger(log)
+}