+++ google2/CHANGELOG.md
@@ -10,5 +10,8 @@ This file documents changes to the `workbench-google2` library, including notes 
 - Add `GoogleStorageService`
 - Add `GooglePubSub`
 - Expose `topicAdminClientResource`
+- Add `GoogleTopicAdmin`, `GoogleStorageNotificationCreater
+- Add `GoogleStorageService.resource` helper for constructing `GoogleStorageService`
+- Add `GoogleStorageService.getObject`
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.1-326ebf4"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.1-TRAVIS-REPLACE-ME"`